🔨 Refactor: 팀 생성, 수정 공통 컴포넌트 적용
